// src/app/features/dashboard/models/add-user-request.model.ts
export interface AddUserRequest {
    Id: string
    firstName: string
    lastName: string
    email: string
    phone: string
    roleId: string
    username: string
    password: string
    permissions: Permission[] 
}

export interface Permission {
    permissionId: string
    isReadable: boolean
    isWritable: boolean
    isDeletable: boolean
}